moved page handlers to different package clean up code
