+++ fyne/pages/handler/sendpagehandler/confirmationwindow.go
@@ -16,8 +16,8 @@ import (
 
 	"github.com/raedahgroup/godcr/fyne/assets"
 	"github.com/raedahgroup/godcr/fyne/layouts"
-	"github.com/raedahgroup/godcr/fyne/pages/constantvalues"
-	"github.com/raedahgroup/godcr/fyne/pages/multipagecomponents"
+	"github.com/raedahgroup/godcr/fyne/pages/handler/constantvalues"
+	"github.com/raedahgroup/godcr/fyne/pages/handler/multipagecomponents"
 	"github.com/raedahgroup/godcr/fyne/widgets"
 )
