+++ fyne/pages/send.go
@@ -2,7 +2,6 @@ package pages
 
 import (
 	"image/color"
-	"log"
 	"sort"
 
 	"fyne.io/fyne"
@@ -12,9 +11,9 @@ import (
 	"github.com/decred/dcrd/dcrutil"
 	"github.com/raedahgroup/dcrlibwallet"
 
-	"github.com/raedahgroup/godcr/fyne/pages/constantvalues"
-	"github.com/raedahgroup/godcr/fyne/pages/multipagecomponents"
-	"github.com/raedahgroup/godcr/fyne/pages/sendpagehandler"
+	"github.com/raedahgroup/godcr/fyne/pages/handler/constantvalues"
+	"github.com/raedahgroup/godcr/fyne/pages/handler/multipagecomponents"
+	"github.com/raedahgroup/godcr/fyne/pages/handler/sendpagehandler"
 	"github.com/raedahgroup/godcr/fyne/widgets"
 )
 
@@ -41,19 +40,18 @@ var sendPage sendPageDynamicData
 func sendPageContent(multiWallet *dcrlibwallet.MultiWallet, window fyne.Window) fyne.CanvasObject {
 	openedWalletIDs := multiWallet.OpenedWalletIDsRaw()
 	if len(openedWalletIDs) == 0 {
-		return widget.NewHBox(widgets.NewHSpacer(10), widget.NewLabelWithStyle("Could not retrieve wallets", fyne.TextAlignCenter, fyne.TextStyle{Bold: true}))
+		return widget.NewHBox(widgets.NewHSpacer(10), widget.NewLabelWithStyle(constantvalues.WalletsErr, fyne.TextAlignCenter, fyne.TextStyle{Bold: true}))
 	}
 	sort.Ints(openedWalletIDs)
 
 	var selectedWallet = multiWallet.WalletWithID(openedWalletIDs[0])
 	if selectedWallet == nil {
-		return widget.NewLabelWithStyle("Unable to load MultiWallet", fyne.TextAlignLeading, fyne.TextStyle{})
+		return widget.NewLabelWithStyle(constantvalues.LoadMultiWalletErr, fyne.TextAlignLeading, fyne.TextStyle{})
 	}
 
 	selectedWalletAccounts, err := selectedWallet.GetAccountsRaw(dcrlibwallet.DefaultRequiredConfirmations)
 	if err != nil {
-		log.Println("Error while getting accounts for wallet", err.Error())
-		return widget.NewLabel("Error while getting accounts for wallet")
+		return widget.NewLabel(constantvalues.AccountDetailsErr)
 	}
 
 	initSendPageDynamicContent(openedWalletIDs, selectedWalletAccounts)
@@ -92,9 +90,10 @@ func sendPageContent(multiWallet *dcrlibwallet.MultiWallet, window fyne.Window) 
 		SelfSending:      sendingToAccountSelectorObjects,
 		Window:           window,
 	}
+
 	err = initSendPage.InitAllSendPageComponents()
 	if err != nil {
-		return widget.NewLabelWithStyle("Unable to load Send Page components, "+err.Error(), fyne.TextAlignLeading, fyne.TextStyle{})
+		return widget.NewLabelWithStyle(constantvalues.SendPageLoadErr, fyne.TextAlignLeading, fyne.TextStyle{})
 	}
 
 	return widget.NewHBox(widgets.NewHSpacer(30), initSendPage.SendPageContents)
